% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{BayesClust}
\alias{BayesClust}
\title{Detect Spatial Functional Clusters Based on Bayesian Spanning Tree}
\usage{
BayesClust(
  data,
  formula = Yk ~ 1 + Xk,
  family = "normal",
  graph0,
  init_val,
  hyperpar,
  correction = F,
  detailed = F,
  MCMC,
  BURNIN,
  THIN,
  path_save,
  seed = 1234,
  ...
)
}
\arguments{
\item{data}{A list containing data matrices 'Y' and 'X', and 'N' for the number of trials or cases.}

\item{formula}{A formula object representing the model to be fitted.}

\item{graph0}{Initial spanning tree used for the Bayesian model.}

\item{init_val}{List of initial values for parameters 'trees', 'beta', and 'cluster'.}

\item{hyperpar}{List containing hyperparameters used in the model.}

\item{MCMC}{Integer, number of MCMC iterations to perform.}

\item{BURNIN}{Integer, number of burn-in iterations to discard.}

\item{THIN}{Integer, thinning interval for recording the results.}

\item{path_save}{Character, the path where results should be saved.}

\item{seed}{Integer, seed value for random number generation, defaults to 1234.}
}
\value{
NULL The function primarily outputs results to a specified path and does not return anything.
}
\description{
This function implements a Reversible Jump Markov Chain Monte Carlo (RJMCMC) algorithm
for detecting spatial functional clusters based on a Bayesian analysis of spanning trees.
It handles various types of moves including birth, death, change, and hyperparameter updates
to explore the space of possible cluster configurations.
}
\examples{
# Example setup (note: actual data and parameters need to be defined)
\dontrun{
data <- list(Y = matrix(rnorm(100), ncol=10), X = matrix(rnorm(100), ncol=10))
formula <- Y ~ X1 + X2
inla.extra <- list(correction = TRUE)
graph0 <- matrix(sample(0:1, 100, replace=TRUE), ncol=10)
init_val <- list(trees = graph0, beta = runif(10), cluster = sample(1:5, 10, replace=TRUE))
hyperpar <- list(c = 0.5)
path_save <- "path/to/save/results/"

BayesClust(data,family, formula, graph0, init_val, hyperpar, MCMC, BURNIN, THIN, path_save, seed = 1234)
}
}
