% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsfc-package.R
\docType{package}
\name{bsfc-package}
\alias{bsfc-package}
\alias{mbsi-package}
\title{bsfc: Bayesian spatial functional clustering index}
\description{
The \code{bsfc} package performs Bayesian spatial functional clustering.
}
\author{
Ruiman Zhong, Erick A. Chacón-Montalván, Paula Moraga
}
