% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_build.R
\name{ConstructGraph0}
\alias{ConstructGraph0}
\title{Construct Initial Graph from Spatial Data}
\usage{
ConstructGraph0(map, method = "knn", para = 10, seed = 1234)
}
\arguments{
\item{map}{An sf object from which coordinates are extracted to construct the graph.}

\item{method}{Character string specifying the method to use for graph construction.
Available methods are 'knn', 'knn_geo', 'rnn', 'rnn_geo', 'deltri', and 'adjmat'.}

\item{para}{Parameter specific to the chosen method; varies by method, such as the number of neighbors or a distance threshold.}

\item{seed}{Integer value to set the random seed for reproducibility.}
}
\value{
An igraph object representing the constructed graph.
}
\description{
This function constructs a graph from spatial coordinates using various methods,
including k-nearest neighbors, geographical k-nearest neighbors, and Delaunay triangulation.
}
\examples{
\dontrun{
  library(sf)
  # Assuming 'map' is an sf object with spatial data
  graph <- ConstructGraph0(map, method = 'knn', para = 3)
  print(graph)
}

}
