% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-cluster.R
\name{continue_bsfc}
\alias{continue_bsfc}
\title{Continue MCMC Clustering Procedure}
\usage{
continue_bsfc(
  result,
  Y,
  X = NULL,
  N = NULL,
  formula = Yk ~ 1 + Xk,
  family = "normal",
  hyperpar = list(c = 0.5),
  correction = FALSE,
  niter = 100,
  burnin = 0,
  thin = 1,
  path_save = NULL,
  ...
)
}
\arguments{
\item{result}{List containing results from a previous BSFC clustering, expected to include
a minimum spanning tree (\code{mst}) and cluster assignments.}

\item{Y}{Numeric vector or matrix of response variables used in the initial clustering.}

\item{X}{Optional numeric vector or matrix of covariates used in the model (default is \code{NULL}).}

\item{N}{Optional numeric vector specifying the number of trials or cases, relevant for
families like binomial (default is \code{NULL}).}

\item{formula}{An object of class \code{\link[stats]{formula}}, specifying the model used in the analysis.}

\item{family}{Character string specifying the family of distributions to use for the model.
Defaults to "normal".}

\item{hyperpar}{List of hyperparameters used in the model; defaults to a list with \code{c = 0.5}.}

\item{correction}{Logical indicating whether correction for overdispersion or other factors
should be applied. Defaults to \code{FALSE}.}

\item{niter}{Integer specifying the number of additional MCMC iterations to perform.}

\item{burnin}{Integer specifying the number of burn-in iterations to discard in this continuation.}

\item{thin}{Integer specifying the thinning interval for recording the results.}

\item{path_save}{Character string specifying the file path to save the continuation results.
If \code{NULL}, results may not be saved to file.}

\item{...}{Additional arguments passed to the underlying \code{bsfc} function.}
}
\value{
The function does not return a value within R but may output results to files if \code{path_save}
is specified. The results include updated cluster assignments and model parameters after
the additional MCMC iterations.
}
\description{
This function continues a Bayesian spatial fusion clustering (BSFC) process based on a previous result.
It is specifically designed to extend the MCMC iterations from a stopping point using the prior state
of cluster assignments and graphical model parameters.
}
\details{
The function takes the last state of the Markov chain from a previous \code{bsfc} execution and
uses it as the starting point for additional MCMC iterations. This is useful for extending
the analysis without restarting the process, thereby saving computational resources and time.
}
\examples{
# Assuming `result` is already obtained from a previous bsfc run:
\dontrun{
continue_bsfc(result, Y = data$response, X = data$covariates, N = data$trials,
              formula = Yk ~  1 + Xk, family = "normal", hyperpar = list(c = 0.1),
              correction = TRUE, niter = 500, burnin = 50, thin = 5,
              path_save = "path/to/continue_results/")
}
}
