% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherfunctions.R
\name{fit.sfclust}
\alias{fit.sfclust}
\title{Fit function for sfclust objects with cluster sorting}
\usage{
fit.sfclust(
  output,
  k = nrow(output$cluster),
  Y,
  X = NULL,
  N = NULL,
  formula,
  family = "normal",
  correction = FALSE,
  detailed = TRUE,
  ...
)
}
\arguments{
\item{output}{An sfclust object containing the clustering results.}

\item{k}{The row of the cluster matrix to use (default is the last row).}

\item{Y}{The response variable.}

\item{X}{Covariates or predictor variables.}

\item{N}{Number of observations or other input (optional).}

\item{formula}{A model formula.}

\item{family}{A family object specifying the distribution (default is "normal").}

\item{correction}{Logical, indicating whether to apply corrections (default is FALSE).}

\item{detailed}{Logical, indicating whether detailed results should be returned (default is TRUE).}

\item{...}{Additional arguments passed to log_mlik_each.}
}
\value{
A list of models, each of class 'sfclustm', fit for each sorted cluster.
}
\description{
This function applies the log_mlik_each function to fit models for each cluster in sfclust.
The clusters are sorted in descending order based on the number of members in each cluster.
}
