% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherfunctions.R
\name{plot.sfclust}
\alias{plot.sfclust}
\title{Plot function for sfclust objects with a conditional legend and log marginal likelihood}
\usage{
\method{plot}{sfclust}(output, map, k = nrow(output$cluster), title = "Estimated Clusters")
}
\arguments{
\item{output}{An sfclust object containing the clustering results.}

\item{map}{An sf object provided by the user.}

\item{k}{The row of the cluster matrix to use for plotting (default is the last row).}

\item{title}{A title for the plot (default is "Estimated Clusters").}
}
\description{
This function plots the map with estimated clusters from an sfclust object and adds a legend if the number of clusters is less than 10.
It also plots the log marginal likelihood.
}
