% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherfunctions.R
\name{plot.sfclustm}
\alias{plot.sfclustm}
\title{Plot function for sfclustm objects}
\usage{
\method{plot}{sfclustm}(x, a = 1, title = "Fitted Values")
}
\arguments{
\item{x}{An sfclustm object containing the models for each cluster.}

\item{a}{The index of the model to plot (default is 1).}

\item{title}{A title for the plot (default is "Fitted Values").}
}
\description{
This function plots the fitted values (mean) from the a-th model in an sfclustm object.
}
