% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotClusterMap2}
\alias{plotClusterMap2}
\title{Plot True and Estimated Cluster Maps}
\usage{
plotClusterMap2(
  cluster_res,
  map,
  height = 8,
  filepath = file.path(path_im, "fun_us_p1.pdf"),
  title = "Estimated clusters",
  fill = "Cluster",
  palette = NULL
)
}
\arguments{
\item{cluster_res}{Numeric vector of estimated cluster assignments for each area in the map.}

\item{map}{An object of class \code{sf} (simple features), representing the spatial framework for the cluster data.}

\item{filepath}{Character string, the path and name of the file where the plot will be saved.}

\item{title}{Character string, title of the plot}
}
\value{
The function saves the plot to a file and returns an invisible copy of the combined plot object.
The primary output is the jpeg file specified by \code{filepath}.
}
\description{
This function generates a side-by-side visualization of true and estimated cluster maps based on spatial data.
The function uses \code{ggplot2} to create the maps and \code{ggpubr} for arranging the plots side by side. The output
is saved as a jpeg file.
}
\details{
The function creates two maps using the simple features data provided in \code{map}. The first map
visualizes the true cluster assignments using different colors, and the second map displays the estimated
cluster assignments. These maps are helpful for comparing the accuracy of cluster predictions against
ground truth in spatial clustering analyses.

The plots are arranged side by side for easy comparison, with the left map showing true clusters and the
right map showing estimated clusters. Both maps are enhanced with \code{theme_bw()} for a clean background
and labeled appropriately.
}
\examples{
# Assuming `clust_res`, `cluster_true` are available and `map` is an sf object:
\dontrun{
plotClusterMap2(clust_res,3, map, "path/to/save/cluster_maps.jpeg")
}
}
