% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotGraph}
\alias{plotGraph}
\title{Plot a Graph on Spatial Data}
\usage{
plotGraph(map, coords, graph0, title, filepath)
}
\arguments{
\item{map}{An sf object representing the spatial data.}

\item{coords}{Matrix of coordinates where each row corresponds to a node in the graph.}

\item{graph0}{A graph object, typically an igraph object, representing the graph to be plotted.}

\item{title}{Character string specifying the title of the plot.}

\item{filepath}{Character string, the name of the file where the plot will be saved.}
}
\value{
A ggplot object representing the graph plotted over the spatial data.
}
\description{
Visualizes a graph over a spatial map using specified coordinates and graph object.
The function uses the ggraph package to create a custom layout for displaying the graph nodes and edges.
}
\examples{
\dontrun{
  library(sf)
  library(igraph)
  library(ggraph)
  # Assuming 'map' is an sf object and 'coords' are extracted from 'map'
  # and 'graph0' is a previously constructed graph object
  plot <- plotGraph(map, coords, graph0, "Sample Graph")
  print(plot)
}

}
