% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotmlikeIter}
\alias{plotmlikeIter}
\title{Plot Log Marginal Likelihood Over Iterations}
\usage{
plotmlikeIter(mod)
}
\arguments{
\item{log_mlike}{Numeric vector containing the log marginal likelihood values at each iteration.}
}
\value{
The function returns a ggplot object representing the line plot of the log marginal
likelihood over iterations. This object can be further modified with additional \code{ggplot2}
layers or printed directly to view the plot.
}
\description{
This function visualizes the log marginal likelihood over iterations using a line plot.
It is designed to help evaluate the convergence and performance of statistical models
over iterative processes.
}
\details{
The function converts the log marginal likelihood data into a data frame and
then uses \code{ggplot2} to plot these values. The x-axis represents the iteration number,
and the y-axis represents the log marginal likelihood at that iteration. This plot can
be especially helpful in monitoring the convergence of Bayesian models or any iterative
statistical modeling that provides log marginal likelihood as an output.
}
\examples{
# Assuming log_mlike is a vector of log marginal likelihood values:
\dontrun{
plot <- plotmlikeIter(restult$log_mlike)
print(plot)
}
}
