% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotMST}
\alias{plotMST}
\title{Plot a Minimum Spanning Tree on Spatial Data}
\usage{
plotMST(map, coords, graph0, cluster, title, height = NULL, filepath = NULL)
}
\arguments{
\item{map}{An sf object representing the spatial data.}

\item{coords}{Matrix of coordinates where each row corresponds to a node in the graph.}

\item{graph0}{A graph object, typically an igraph object, that includes the MST to be plotted.}

\item{cluster}{A vector indicating the cluster assignment for each node in the graph.}

\item{title}{Character string specifying the title of the plot.}

\item{filepath}{Character string, the name of the file where the plot will be saved.}
}
\value{
A ggplot object representing the MST plotted over the spatial data with nodes colored by cluster.
}
\description{
Visualizes a minimum spanning tree (MST) over a spatial map using specified coordinates and graph object.
Nodes are colored by cluster assignments. The function uses the ggraph package to create a custom layout
for displaying the MST with nodes and edges.
}
\examples{
\dontrun{
  library(sf)
  library(igraph)
  library(ggraph)
  # Assuming 'map' is an sf object, 'coords' are coordinates, 'graph0' is a graph object,
  # and 'cluster' is a vector of cluster assignments
  plot <- plotMST(map, coords, graph0, cluster, "Sample MST")
  print(plot)
}

}
