% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherfunctions.R
\name{print.sfclust}
\alias{print.sfclust}
\title{Print method for sfclust objects}
\usage{
\method{print}{sfclust}(x, ...)
}
\arguments{
\item{x}{An object of class 'sfclust'.}

\item{...}{Additional arguments (currently unused).}
}
\description{
This function prints details of an sfclust object, including the formula, counts, hyperparameter q, and the final log marginal likelihood.
}
