% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherfunctions.R
\name{summary.sfclust}
\alias{summary.sfclust}
\title{Summary method for sfclust objects}
\usage{
\method{summary}{sfclust}(x, a = nrow(x$cluster), ...)
}
\arguments{
\item{x}{An object of class 'sfclust'.}

\item{a}{An integer specifying the row of the cluster matrix to summarize (default is the last row).}

\item{...}{Additional arguments (currently unused).}
}
\description{
This function summarizes the cluster assignments from the a-th row of the cluster matrix.
}
