% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherfunctions.R
\name{summary.sfclustm}
\alias{summary.sfclustm}
\title{Summary method for sfclustm objects}
\usage{
\method{summary}{sfclustm}(x, ...)
}
\arguments{
\item{x}{An object of class 'sfclustm'.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A list of summaries for each model.
}
\description{
This function provides a summary of each model within an sfclustm object.
}
