% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{us_covid_clean}
\alias{us_covid_clean}
\title{US COVID-19 Cleaned Data}
\format{
A list with the following components:
\describe{
\item{Y}{Matrix of COVID-19 case counts.}
\item{N}{Matrix of population data.}
\item{map}{Spatial map of the regions.}
\item{E}{Matrix of expected cases data.}
}
}
\source{
Processed data from the US COVID-19 data
}
\usage{
us_covid_clean
}
\description{
This dataset contains cleaned COVID-19 case counts and population data for different regions in the United States.
}
\keyword{datasets}
