% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalLog.ratio.R
\name{evalLogLike.ratio}
\alias{evalLogLike.ratio}
\title{Evaluate Log-Likelihood Ratio for a Proposed Move}
\usage{
evalLogLike.ratio(
  move,
  log_like_vec,
  control.move,
  Y,
  membership,
  X,
  N,
  family,
  formula,
  correction,
  detailed = F,
  ...
)
}
\arguments{
\item{move}{A character string specifying the type of move; either "split" or "merge".}

\item{log_like_vec}{A numeric vector of existing log likelihoods for all clusters.}

\item{control.move}{A list containing control parameters and information about the move,
such as \code{clust_old}, \code{k}, \code{cluster_newid}, and \code{cluster_rm}.}

\item{Y}{Numeric matrix of data observations.}

\item{membership}{Integer vector indicating the cluster membership of each observation.}

\item{X}{Numeric matrix representing the model's design matrix or covariates.}

\item{N}{The number of trials or cases, applicable in certain statistical families like binomial.}

\item{family}{The family of distributions to use, e.g., "normal", "poisson", etc.}

\item{formula}{An object of class \code{\link[stats]{formula}} specifying the model to be used in INLA.}

\item{correction}{Logical, specifying whether correction calculations should be executed.}

\item{detailed}{Logical, indicating whether detailed output from INLA is required.}
}
\value{
A list containing the updated log likelihood vector and the log likelihood ratio (\code{llratio}).
}
\description{
This function calculates the log-likelihood ratio for a proposed "split" or "merge"
move within a clustering or partitioning algorithm using the Integrated Nested
Laplace Approximation (INLA) approach. It updates local likelihoods based on
whether a split or merge operation has been proposed and supports multiple family
distributions.
}
