% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise-cluster.R
\name{initial_cluster}
\alias{initial_cluster}
\title{Initialize clusters for spatial clustering}
\usage{
initial_cluster(x, nclust = 10, weights = NULL)
}
\arguments{
\item{x}{An \code{sf} or \code{sfc} object representing the spatial polygonal data. Additionally,
it can also be a \code{matrix} or \code{Matrix} object with non-zero values representing the
weighted connectivity between units.}

\item{nclust}{Integer, the initial number of clusters.}

\item{weights}{Optional \code{numeric} vector or \code{matrix} of weights between units in \code{x}.
It should be of dimension \code{n^2}, where \code{n} is the number of units in \code{x}. If NULL,
random weights are assigned.}
}
\value{
A list with three elements: 'graph' containing the initial undirected graph
object, 'mst' containing the initial minimum spanning tree, and 'cluster' containing
the membership of \code{x}.
}
\description{
Creates a undirected graph from spatial polygonal data, computes its minimum spanning
tree (MST) and initialise the \code{nclust} clusters.
}
\examples{

  librar(sf)
  x <- st_make_grid(cellsize = c(1, 1), offset = c(0, 0), n = c(3, 2))
  cluster_ini <- initial_cluster(x, nclust = 3, weights = 1:6))
  print(cluster_ini)

}
